pointerevents/pointerevent_releasePointerCapture_events_to_original_target.html

<!doctype html>
<html>
    <head>
        <title>Pointer Event: releasePointerCapture() - subsequent events follow normal hitting testing mehcanisms</title>
        <meta content="text/html; charset=UTF-8" http-equiv="Content-Type"/>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/"/>
        <meta name="assert" content="After invoking the releasePointerCapture method on an element, subsequent events for the specified pointer must follow normal hit testing mechanisms for determining the event target"/>

        <!-- http://w3c-test.org/resources/testharness.js -->
	    <script src="http://w3c-test.org/resources/testharness.js"></script>
		<!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>
        <script type="text/javascript">

            setup({ explicit_done: true });

            var detected_pointertypes = {};
            add_completion_callback(showPointerTypes);

            var target0;
            var listener;
            var captured_event;
            var f_gotPointerCapture = false;
            var f_lostPointerCapture = false;
            var f_testComplete = false;

            var test_pointerCapture

            function listenerEventHandler(event) {

                if (event.type == "gotpointercapture") {
                    f_gotPointerCapture = true;
                }
                else if (event.type == "lostpointercapture") {
                    f_lostPointerCapture = true;
                }

                if (event.pointerId == captured_event.pointerId && event.type !== "lostpointercapture") {
                    if (!f_lostPointerCapture) {
                        // on first event received for capture, release capture
                        listener.releasePointerCapture(event.pointerId);

                    }
                    else {
                        // if any other events are received, then the test fails
                        test(function () {
                            assert_true(false, event.target.id + "-" + event.type + " should be handled by target element handler");
                        }, "Fail if event received by the listener after lostpointercapture");
                    }
                }
            }

            function targetEventHandler(event) {

                if (!f_gotPointerCapture && event.type == "pointerover") {
                    // pointerover event received will be used to capture events.
                    listener.setPointerCapture(event.pointerId);
                    captured_event = event;
                }

                if (f_lostPointerCapture) {
                    test_pointerCapture.step(function () {
                        assert_equals(event.pointerId, captured_event.pointerId, "pointerID is same for event captured and after release");
                    });
                    test_pointerCapture.done(); // complete test
                }
            }

            function run() {
                listener = document.getElementById("listener");
                target0 = document.getElementById("target0");
                target0.style["touchAction"] = "none";

                test_pointerCapture = async_test("lostpointercapture: subsequent events to target.");  // set up test harness

                // target0 and listener - handle all events               
                for (var i = 0; i < All_Pointer_Events.length; i++) {
                    on_event(target0, All_Pointer_Events[i], targetEventHandler);
                    on_event(listener, All_Pointer_Events[i], listenerEventHandler);
                }
            }

            function showPointerTypes() {
                var complete_notice = document.getElementById("complete-notice");
                var pointertype_log = document.getElementById("pointertype-log");
                var pointertypes = Object.keys(detected_pointertypes); pointertype_log.innerHTML = pointertypes.length ? pointertypes.join(",") : "(none)";
                if (test_pointerCapture.status == test_pointerCapture.NOTRUN
                    || test_pointerCapture.status == test_pointerCapture.TIMEOUT) {
                    complete_notice.innerHTML = "Test FAIL due to time out.  This is due to the browser not supporting pointer, or the manual steps were not run to completion.";
                    complete_notice.style.background = "red";
                }
                complete_notice.style.display = "block";
            }
        </script>

        <style>
            div {
                margin: 0em;
                padding: 2em;
            }

            #target0 {
                background: yellow;
                border: 1px solid orange;
            }
            #listener {
                background: orange;
                border: 1px solid orange;
                position: absolute;
                top: -100px;
            }

            #complete-notice {
                border: 1px solid #0a0;
                display: none;
            }

            #pointertype-log {
                font-weight: bold;
            }
        </style>
    </head>
    <body onload="run()">
        <div id="listener"></div>
        <h1>Pointer Event: releasePointerCapture() - subsequent events follow normal hitting testing mehcanisms</h1>
        <h4> 
            Test Description: 
            After invoking the releasePointerCapture method on an element, subsequent events for the specified 
            pointer must follow normal hit testing mechanisms for determining the event target 
        </h4>
        <br />
        <div id="target0">
            Use mouse, touch or pen to contact here and move around.
        </div>
        
        <div id="complete-notice">
            <p>Test complete:  Scroll to Summary to view Pass/Fail Results.</p>
            <p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
            <p>Would you like to <a href="#">run the tests again</a> with a different pointer type?</p>
        </div>

        <div id="log"></div>
    </body>
</html>
